feature: use data type from signalk-schema
Use @signalk/signalk-schema.getUnits to derive data type in InfluxDb:
if the schema has unit for the path and the unit is not a timestamp
force numeric data type in Influxdb.

This will prevent value type related errors where the first value that
is written to Influx has a bogus data type and all subsequent values
just create errors because the type that ended in Influx is off.

Fixes #7.